# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Wisnu API",
        "description": "API documentation for Hotel Wisnu Management System",
        "contact": {
            "name": "Hotel Wisnu API Support",
            "email": "admin@hotelwisnu.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Hotel Wisnu API Server"
        }
    ],
    "paths": {
...
        "/room-types": {
            "get": {
                "tags": [
                    "Room Types"
                ],
                "summary": "Get all room types",
                "description": "Retrieve a list of all room types",
                "operationId": "getRoomTypes",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "room_type_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "type_name": {
                                                "type": "string",
                                                "example": "Deluxe Room"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Spacious room with king size bed"
                                            },
                                            "base_price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 1500
                                            },
                                            "max_occupancy": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "amenities": {
                                                "type": "string",
                                                "example": "WiFi, TV, Mini Bar"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Room Types"
                ],
                "summary": "Create a new room type",
                "description": "Create a new room type with the provided data",
                "operationId": "storeRoomType",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type_name",
                                    "description",
                                    "base_price",
                                    "max_occupancy",
                                    "amenities"
                                ],
                                "properties": {
                                    "type_name": {
                                        "type": "string",
                                        "example": "Deluxe Room"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Spacious room with king size bed"
                                    },
                                    "base_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500
                                    },
                                    "max_occupancy": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "amenities": {
                                        "type": "string",
                                        "example": "WiFi, TV, Mini Bar"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Room type created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Room type created successfully"
                                        },
                                        "room_type": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}